pyproject.toml
[project]
name = "hass-livebox-component"
version = "0.0.0"
description = "Development environment for the Orange Livebox Home Assistant custom component"
requires-python = ">=3.13.2"
dependencies = []

[dependency-groups]
dev = [
  "aiosysbus",
  "async-upnp-client",
  "pre-commit",
  "pycares<5",
  "pyrefly",
  "pylint",
  "pytest-homeassistant-custom-component==0.13.272",
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = ["tests/"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
log_cli = true

[tool.uv]
package = false

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["UP040"]

[tool.pyrefly]
project-includes = [
    "**/*.py*",
    "**/*.ipynb",
]
